Fix shellcheck warning across the whole testsuite
Shellcheck is becoming a standard for posting patches upstream. For
this reason fix all of the warnings before the next set of patches will
be sent upstream.

+++ base_buildid/cleanup.sh
@@ -13,19 +13,19 @@
 touch $CURRENT_TEST_DIR/BUILDIDDIRS
 while read line; do
 	BUILDIDDIR="$line"
-	if [[ " ${BUILDIDDIR}/*" =~ ' /*' ]] || [[ "$BUILDIDDIR" =~ /$ ]]; then
+	if [[ " ${BUILDIDDIR}/*" == *' /*'* ]] || [[ "$BUILDIDDIR" =~ /$ ]]; then
 		true # skipping deletion
 	else
 		rm -rf $BUILDIDDIR/.b*
-		rm -rf $BUILDIDDIR/*
+		rm -rf ${BUILDIDDIR:?}/*
 		rmdir $BUILDIDDIR 2> /dev/null
 	fi
 done < $CURRENT_TEST_DIR/BUILDIDDIRS
 rm -f $CURRENT_TEST_DIR/BUILDIDDIRS
 
 if [ ! -n "$PERFSUITE_RUN_DIR" ]; then
-	find . -name \*.log | xargs -r rm
-	find . -name \*.err | xargs -r rm
+	find . -name \*.log -print0 | xargs -r -0 rm
+	find . -name \*.err -print0 | xargs -r -0 rm
 	test -e perf.data && rm -rf perf.data
 	test -e perf.data.old && rm -rf perf.data.old
 	test -e perfnew.data && rm -rf perfnew.data
